fix: Restore frontend lint budget
Keep factory intake hooks and fixture routes in focused modules so the
frontend stays within its file-size budget. Stabilize the settings save
callback to satisfy the JSX binding rule.
