+++ web_src/src/pages/factories/__fixtures__/handlers.ts
@@ -1,4 +1,5 @@
 import { EMPTY_USAGE_REPORT } from "./usageReportFixtures";
+import { factoryIntakeRoutes } from "./factoryIntakeHandlers";
 import {
   defaultFactoriesFixture,
   ORGANIZATION_USERS,
@@ -12,11 +13,9 @@ import { DEFAULT_ARTIFACTS_BY_ORDER_ID, DEFAULT_EVENTS_BY_ORDER_ID } from "./fac
 import { DEFAULT_CHECKS_BY_ORDER_ID } from "./workOrderCheckFixtures";
 import type {
   FactoriesFactory,
-  FactoriesFactoryIntake,
   FactoriesFactoryLine,
   FactoriesFactoryOnboarding,
   FactoriesUpdateFactoryOnboardingBody,
-  FactoryIntakeSettings,
   FactoriesWorkOrder,
   FactoriesWorkOrderEvent,
   FactoriesWorkOrderLineDispatch,
@@ -29,10 +28,6 @@ import { isValidWorkspaceKey, suggestWorkspaceKeyFromName, WORKSPACE_KEY_MAX_LEN
 import { metricsForLine } from "../pages/lineListMetricsMockData";
 
 export type { FactoriesFixture };
-
-export const factoryPageIds = {
-  organizationId: defaultFactoriesFixture.organizationId,
-};
 
 const re = (pattern: string): RegExp => new RegExp(`^${pattern}$`);
 
@@ -184,83 +179,6 @@ function factoryDetailRoutes(fixture: FactoriesFixture): FactoriesRoute[] {
     {
       pattern: re("/api/v1/factories/([^/]+)/usage"),
       resolve: (match) => ({ json: fixture.usageByFactoryId?.[match[1]] ?? EMPTY_USAGE_REPORT }),
-    },
-  ];
-}
-
-const INTAKE_NAME_BY_SOURCE: Record<string, string> = {
-  SOURCE_GITHUB_ISSUES: "GitHub issues",
-  SOURCE_SENTRY_EXCEPTIONS: "Sentry exceptions",
-  SOURCE_PAGERDUTY_INCIDENTS: "PagerDuty incidents",
-};
-
-function factoryIntakes(fixture: FactoriesFixture, factoryId: string): FactoriesFactoryIntake[] {
-  const existing = fixture.intakesByFactoryId?.[factoryId];
-  if (existing) {
-    return existing;
-  }
-  const created: FactoriesFactoryIntake[] = [];
-  fixture.intakesByFactoryId = { ...fixture.intakesByFactoryId, [factoryId]: created };
-  return created;
-}
-
-function factoryIntakeRoutes(fixture: FactoriesFixture): FactoriesRoute[] {
-  return [
-    {
-      pattern: re("/api/v1/factories/([^/]+)/intakes/([^/]+)/runs"),
-      resolve: (match) => ({ json: { runs: fixture.intakeRunsByIntakeId?.[match[2]] ?? [] } }),
-    },
-    {
-      pattern: re("/api/v1/factories/([^/]+)/intakes/([^/]+)"),
-      resolve: (match, method, body) => {
-        const intakes = factoryIntakes(fixture, match[1]);
-        const index = intakes.findIndex((intake) => intake.id === match[2]);
-        if (index < 0) {
-          return { json: {} };
-        }
-        if (method === "DELETE") {
-          intakes.splice(index, 1);
-          return { json: {} };
-        }
-        if (method !== "PATCH" && method !== "PUT") {
-          return null;
-        }
-        const request = (body ?? {}) as { name?: unknown; settings?: FactoryIntakeSettings };
-        const updated: FactoriesFactoryIntake = {
-          ...intakes[index],
-          ...(stringOrEmpty(request.name) ? { name: stringOrEmpty(request.name) } : {}),
-          ...(request.settings ? { settings: { ...intakes[index].settings, ...request.settings } } : {}),
-        };
-        intakes[index] = updated;
-        return { json: { intake: updated } };
-      },
-    },
-    {
-      pattern: re("/api/v1/factories/([^/]+)/intakes"),
-      resolve: (match, method, body) => {
-        const intakes = factoryIntakes(fixture, match[1]);
-        if (method !== "POST") {
-          return { json: { intakes } };
-        }
-        const request = (body ?? {}) as { source?: unknown; name?: unknown; confidencePct?: unknown };
-        const source = stringOrEmpty(request.source) || "SOURCE_GITHUB_ISSUES";
-        const created: FactoriesFactoryIntake = {
-          id: `storybook-intake-${intakes.length + 1}`,
-          canvasId: `storybook-intake-canvas-${intakes.length + 1}`,
-          factoryId: match[1],
-          name: stringOrEmpty(request.name) || INTAKE_NAME_BY_SOURCE[source] || source,
-          source: source as FactoriesFactoryIntake["source"],
-          settings: {
-            confidencePct: typeof request.confidencePct === "number" ? request.confidencePct : 65,
-            labels: [],
-            labelFilterMode: "LABEL_FILTER_MODE_INCLUDE",
-            assignment: "ASSIGNMENT_ANY",
-          },
-          healthy: true,
-        };
-        intakes.push(created);
-        return { json: { intake: created } };
-      },
     },
   ];
 }
