+++ web_src/src/pages/factories/pages/useIntakeAutomationRuns.ts
@@ -1,4 +1,4 @@
-import { useFactoryIntakeRuns } from "@/hooks/useFactoryData";
+import { useFactoryIntakeRuns } from "@/hooks/useFactoryIntakeData";
 import { useMemo } from "react";
 
 import { intakeRunsFromApi, type IntakeAutomationRun } from "./intakeSourceSettingsModel";
